Status and types tables dropped Imported from Changeset 2567

+++ GaneEcommerce.Services/Core/Finance/Financials.cs
@@ -20,10 +20,10 @@ namespace Ganedata.Core.Services
 
             var context = DependencyResolver.Current.GetService<IApplicationContext>();
 
-            var totalInvoiced = context.AccountTransactions.Where(x => x.AccountId == accountId && (x.AccountTransactionTypeId == (int)AccountTransactionTypeEnum.InvoicedToAccount) && x.IsDeleted != true).ToList().Sum(m => m.Amount);
+            var totalInvoiced = context.AccountTransactions.Where(x => x.AccountId == accountId && (x.AccountTransactionTypeId == AccountTransactionTypeEnum.InvoicedToAccount) && x.IsDeleted != true).ToList().Sum(m => m.Amount);
 
-            var totalPaid = context.AccountTransactions.Where(x => x.AccountId == accountId && (x.AccountTransactionTypeId == (int)AccountTransactionTypeEnum.PaidByAccount
-            || x.AccountTransactionTypeId == (int)AccountTransactionTypeEnum.Refund || x.AccountTransactionTypeId == (int)AccountTransactionTypeEnum.Discount || x.AccountTransactionTypeId == (int)AccountTransactionTypeEnum.CreditNote) && x.IsDeleted != true).ToList().Sum(m => m.Amount);
+            var totalPaid = context.AccountTransactions.Where(x => x.AccountId == accountId && (x.AccountTransactionTypeId == AccountTransactionTypeEnum.PaidByAccount
+            || x.AccountTransactionTypeId == AccountTransactionTypeEnum.Refund || x.AccountTransactionTypeId == AccountTransactionTypeEnum.Discount || x.AccountTransactionTypeId == AccountTransactionTypeEnum.CreditNote) && x.IsDeleted != true).ToList().Sum(m => m.Amount);
 
             balance = totalInvoiced - totalPaid;
